Issue #152

* Implement delete multiple files
* Fix #217 

See merge request !248

+++ src/com/teamgamma/musicmanagementsystem/model/SongManager.java
@@ -38,8 +38,6 @@ public class SongManager {
     private List<Item> m_itemsToMove;
     private File m_copyDest;
     private File m_moveDest;
-    private File m_addedFile;
-    private File m_deletedFile;
     private File m_renamedFile;
 
     // For observer pattern
@@ -83,8 +81,6 @@ public class SongManager {
         m_itemsToMove = null;
         m_copyDest = null;
         m_moveDest = null;
-        m_addedFile = null;
-        m_deletedFile = null;
         m_renamedFile = null;
 
         m_selectedCenterFolder = null;
@@ -203,6 +199,18 @@ public class SongManager {
     }
 
     /**
+     * Remove itemsToSkip from the list of items
+     *
+     * @param items         list of items
+     * @param itemsToSkip   list of items to remove from items
+     */
+    private void removeItemsToSkip(List<Item> items, List<Item> itemsToSkip) {
+        for (Item itemToSkip: itemsToSkip) {
+            items.remove(itemToSkip);
+        }
+    }
+
+    /**
      * Copy files in buffer to destination
      *
      * @param dest the destination folder
@@ -215,11 +223,23 @@ public class SongManager {
             throw new Exception("Files to copy should not be null");
         }
 
+        List<Item> itemsToSkip = new ArrayList<>();
         for (Item itemToCopy : m_itemsToCopy) {
-            if (!FileManager.copyFilesRecursively(itemToCopy.getFile(), dest)) {
-                throw new IOException("Fail to copy");
+            try {
+                if (dest.getAbsolutePath().equals(itemToCopy.getFile().getParent())){
+                    return;
+                }
+
+                if (!FileManager.copyFilesRecursively(itemToCopy.getFile(), dest)) {
+                    throw new IOException("Fail to copy");
+                }
+            } catch (FileAlreadyExistsException ex) {
+                System.out.println("### Skipping song: " + itemToCopy.getFile());
+                itemsToSkip.add(itemToCopy);
             }
         }
+
+        removeItemsToSkip(m_itemsToCopy, itemsToSkip);
 
         m_copyDest = dest;
 
@@ -242,9 +262,21 @@ public class SongManager {
             throw new Exception("Files to move should not be null");
         }
 
+        List<Item> itemsToSkip = new ArrayList<>();
         for(Item itemToMove : m_itemsToMove) {
-            FileManager.moveFile(itemToMove.getFile(), destDir);
+            try {
+                if (itemToMove.getFile().getParent().equals(destDir.getAbsolutePath())) {
+                    return;
+                }
+
+                FileManager.moveFile(itemToMove.getFile(), destDir);
+            } catch (FileAlreadyExistsException ex) {
+                System.out.println("### Skipping song: " + itemToMove.getFile());
+                itemsToSkip.add(itemToMove);
+            }
         }
+
+        removeItemsToSkip(m_itemsToMove, itemsToSkip);
 
         m_moveDest = destDir;
 
@@ -258,31 +290,34 @@ public class SongManager {
     /**
      * Delete a file
      *
-     * @param fileToDelete
+     * @param filesToDelete
      * @throws Exception
      */
-    public void deleteFile(File fileToDelete) throws Exception {
-        if (m_rightFolderSelected != null && m_rightFolderSelected.getAbsolutePath().equals(fileToDelete.getAbsolutePath())) {
-            m_rightFolderSelected = null;
-        }
-        if (m_selectedCenterFolder != null && m_selectedCenterFolder.getAbsolutePath().equals(fileToDelete.getAbsolutePath())) {
-            m_selectedCenterFolder = null;
-        }
+    public void deleteFile(List<File> filesToDelete) throws Exception {
+        FileActions deleteFileAction = new ConcreteFileActions();
+        for (File fileToDelete: filesToDelete) {
+            // Skip this song if it does not exist in the file system
+            if (!fileToDelete.exists()) {
+                break;
+            }
 
-        if (!FileManager.removeFile(fileToDelete)) {
-            throw new FileSystemException("File " + fileToDelete.getAbsolutePath() + " could not be deleted");
+            if (m_rightFolderSelected != null && m_rightFolderSelected.getAbsolutePath().equals(fileToDelete.getAbsolutePath())) {
+                m_rightFolderSelected = null;
+            }
+            if (m_selectedCenterFolder != null && m_selectedCenterFolder.getAbsolutePath().equals(fileToDelete.getAbsolutePath())) {
+                m_selectedCenterFolder = null;
+            }
+
+            if (!FileManager.removeFile(fileToDelete)) {
+                throw new FileSystemException("File " + fileToDelete.getAbsolutePath() + " could not be deleted");
+            }
+
+            deleteFileAction.add(Action.DELETE, fileToDelete);
         }
-
-        m_deletedFile = fileToDelete;
-
-        FileActions deleteFileAction = new ConcreteFileActions(Action.DELETE, fileToDelete);
 
         updateFilesInFileTree(deleteFileAction);
 
         notifyFileObservers(deleteFileAction);
-
-        // Clear file to delete buffer
-        m_deletedFile = null;
     }
 
     /**
@@ -599,14 +634,6 @@ public class SongManager {
     /**********
      * Getters and setters
      *************/
-
-    public File getM_addedFile(){
-        return m_addedFile;
-    }
-
-    public File getM_deletedFile() {
-        return m_deletedFile;
-    }
 
     public File getM_renamedFile() {
         return m_renamedFile;
